requirements.txt
flask>=2.0.0
pandas>=1.3.0
numpy>=1.20.0
sqlalchemy>=1.4.0
pymysql>=1.0.2
torch>=1.9.0
scikit-learn>=0.24.0
ccxt>=1.60.0
websocket-client>=1.1.0
requests>=2.26.0
python-dateutil>=2.8.2
matplotlib>=3.4.0
plotly>=5.3.0
psutil>=5.8.0
talib-binary>=0.4.0  # این پکیج ممکن است نیاز به نصب جداگانه داشته باشد

python main.py --config config.json
python main.py --config config.json